detect concurrent modification
This change effectively makes all modifications to Set and Map -- the
only two mutable types in this package -- atomic and resilient to
concurrent modification. This should add to the safety of using them
in goroutines or, at least, help you catch places where you write from
multiple goroutines.

This doesn't solve every problem. For example, there is no transaction
to ensure that if you read and then write that there is no race
between them. You still need to design your program to only modify in
a single goroutine or use syncronization around reads and writes to
ensure isolation.

The unsafe and atomic packages have all kinds of warnings on them but
I think I got this right. I kept their use to a very specific spot in
the code so that it should be easy to inspect.