change from error to succeeded bool
Returning an error probably encourages checking at any cost. (Always
check your errors, right?). In many cases, though, callers not really
care if it succeeds or not as long as the structure wasn't changed in
the event of failure. I think returning a bool like this sends more of
a "you can check me if you want but don't feel obligated" kind of
message.