optimize when left == right
If left and right are the same pointer, we know they point to exactly
the same structure and there will never be any difference.

This should make a huge difference if a large trie is modified (old ->
new) in a small way and then you run diff. Most of the old and new
structures will be common between the two.