simplify LongestMatch API
I agonized over this a bit. At first, I wanted to return the Match
value indicating the type of match. The more I looked at it though,
the more I wanted to just stick with a bool. Here are some reasons for
this thinking.

1. This change makes it more like how go map works.
2. If you want an exact match, use Get.
3. If you are using LongestMatch, chances are you probably don't care
   if it is exact or not. In that case, the following feels more
   natural to me than comparing the result against an enum value:

   if found { ... } // (as opposed to if matched != MatchNone)

4. If you do care whether it is exact or not, it is pretty natural to
   write the following. To me, this reads as well or better than
   having to always compare against MatchExact.

   if found && prefix == searchPrefix { ... }

There may be cases where the old way is preferable but I think overall
this is the better way.