find optimal prefix to allocate to avoid fragmentation
Given two sets, the first one representing the full IP space which can
be used, and a second representing what has already been used or
allocated, this algorthim can find a best fit prefix that you can add to
reserved set which will minimize the fragmentation of the resulting
reserved set.

This can be used to allocate IP address subnets in the best way possible
to avoid situations where a subnet cannot be allocated because the space
was not allocated carefully enough.

It cannot prevent all fragmentation. Since it has no knowledge of how
prefixes may be deallocated in the future, it cannot make decisions to
allocate so that deallocations will open up the most contiguous space
possible.